Better evaluation during training! We now compare free energy on training data against some validation data as a way to monitor overfitting. Also fixed up score_samples, and changed its output to be more interpretable (to me).
Some refactoring/cleanup also, but lots remains to be done.
